// planner-web/src/api/generated/model/financialPlanResponseDto.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Financial Planner API
 * Versioned API for normalized financial plans
 * OpenAPI spec version: 1.0
 */
import type { FinancialPlanResponseDtoStatus } from "./financialPlanResponseDtoStatus"

export interface FinancialPlanResponseDto {
  id: string
  metadataId: string
  schemaVersion: string
  name: string
  currency: string
  startDate: string
  /** @nullable */
  endDate: string | null
  status: FinancialPlanResponseDtoStatus
  /** @nullable */
  objective: string | null
  /** @nullable */
  projectedDebtFreeDate: string | null
  /** @nullable */
  projectedEmergencyFund: number | null
  createdAt: string
  updatedAt: string
}